move shared ini file to singleton
