+++ lib/credentials/shared_ini_file_credentials.js
@@ -1,5 +1,5 @@
 var AWS = require('../core');
-var sharedIniFile = require('../shared-ini-file');
+var sharedIniFile = require('../shared-ini');
 var STS = require('../../clients/sts');
 
 /**
