+++ lib/node_loader.js
@@ -10,9 +10,6 @@ util.querystring = require('querystring');
 util.environment = 'nodejs';
 util.createEventStream = util.stream.Readable ?
   require('./event-stream/streaming-create-event-stream').createEventStream : require('./event-stream/buffered-create-event-stream').createEventStream;
-
-require('./shared-ini-file');
-
 util.realClock = require('./realclock/nodeClock');
 util.clientSideMonitoring = {
   Publisher: require('./publisher').Publisher,
@@ -46,8 +43,6 @@ require('./credentials/ecs_credentials');
 require('./credentials/environment_credentials');
 require('./credentials/file_system_credentials');
 require('./credentials/shared_ini_file_credentials');
-
-// var SharedIniFile = require('./shared_ini');
 
 // Setup default chain providers
 // If this changes, please update documentation for
@@ -93,7 +88,7 @@ AWS.util.update(AWS.Config.prototype.keys, {
         {isConfig: true, filename: env[AWS.util.sharedConfigFileEnv]}
       ];
       while (!region && toCheck.length) {
-        var sharedIniFile = require('./shared-ini-file');
+        var sharedIniFile = require('./shared-ini');
         var configFile = sharedIniFile.loadFrom(toCheck.shift());
         var profile = configFile[env.AWS_PROFILE || AWS.util.defaultProfile];
         region = profile && profile.region;
