+++ lib/publisher/configuration.js
@@ -1,5 +1,5 @@
 var AWS = require('../core');
-var sharedIniFile = require('../shared-ini-file');
+var sharedIniFile = require('../shared-ini');
 /**
  * Resolve client-side monitoring configuration from either environmental variables
  * or shared config file. Configurations from environmental variables have higher priority
