--- lib/shared-ini-file/index.d.ts
@@ -1,12 +1,0 @@
-export interface LoadFileOptions {
-  filename?: string,
-  isConfig?: boolean,
-}
-
-export interface IniFileContent {
-  [key: string]: {[key: string]: string}
-}
-
-export function loadFrom(options: LoadFileOptions): IniFileContent;
-
-export function clearCachedFiles(): void;
