--- lib/shared-ini-file/index.js
@@ -1,2 +1,0 @@
-var SharedIniFile = require('./shared_ini');
-module.exports = new SharedIniFile();
