+++ lib/shared-ini/shared_ini.js
@@ -8,17 +8,6 @@ var path = require('path');
 module.exports = AWS.util.inherit({
   constructor: function SharedIniFile() {
     this.resolvedProfiles = {};
-    //ignore exceptions if defaults cannot be loaded
-    try {
-      var defaults = {};
-      var isConfig = true;
-      var configFilePath = this.getDefaultFilePath(isConfig);
-      defaults[configFilePath] = this.parseFile(configFilePath, isConfig);
-      isConfig = false
-      var credentialFilePath = this.getDefaultFilePath(isConfig);
-      defaults[credentialFilePath] = this.parseFile(credentialFilePath, isConfig);
-      this.resolvedProfiles = defaults;
-    } catch (e) {}
   },
 
   clearCachedFiles: function clearCachedFiles() {
