+++ test/config.spec.js
@@ -1,6 +1,7 @@
 var hasProp = {}.hasOwnProperty;
 var helpers = require('./helpers');
 var AWS = helpers.AWS;
+var sharedIniFile = require('../lib/shared-ini');
 var configure = function(options) {
   return new AWS.Config(options);
 };
@@ -40,6 +41,7 @@ describe('AWS.Config', function() {
 
     afterEach(function() {
       process.env = oldEnv;
+      sharedIniFile.clearCachedFiles();
     });
 
     it('defaults to undefined', function() {
