+++ test/helpers.js
@@ -53,7 +53,6 @@
   spies = null;
 
   beforeEach(function() {
-    if (AWS.util.isNode()) require('../lib/shared-ini-file').clearCachedFiles();
     return spies = [];
   });
