+++ test/shared_ini.spec.js
@@ -1,5 +1,5 @@
 var helpers = require('./helpers');
-var SharedIniFile = require('../lib/shared-ini-file/shared_ini');
+var SharedIniFile = require('../lib/shared-ini/shared_ini');
 var os = require('os');
 var AWS = helpers.AWS;
 
@@ -27,75 +27,11 @@ if (AWS.util.isNode()) {
     });
   
     describe ('constructor', function() {
-      it('skip loading defaults files if cannot be loaded', function() {
+      it('should only read config file lazily when need to', function() {
         helpers.spyOn(os, 'homedir').andReturn('/foo/bar/baz');
-        helpers.spyOn(AWS.util, 'readFileSync').andCallThrough();
-        var error;
-        var sharedIniFile;
-        try {
-          sharedIniFile = new SharedIniFile();
-        } catch (e) {
-          error = e;
-        }
-        expect(AWS.util.readFileSync.calls.length).to.equal(1);
-        expect(error).to.eql(undefined);
-        //if the path is not cached, loadFrom() will try to read file again
-        try {
-          sharedIniFile.loadFrom({filename: '/foo/bar/baz/.aws/credentials'});
-        } catch (e) {
-          error = e;
-        }
-        expect(AWS.util.readFileSync.calls.length).to.equal(2);
-        expect(error).not.to.undefined;
-      });
-  
-      it('can load config and credentials by default if available', function() {
-        helpers.spyOn(os, 'homedir').andReturn('/foo/bar/baz');
-        helpers.spyOn(AWS.util, 'readFileSync').andReturn('[default]\naws_access_key_id = akid\naws_secret_access_key = secret\naws_session_token = session');
+        helpers.spyOn(AWS.util, 'readFileSync');
         var sharedIniFile = new SharedIniFile();
-        expect(AWS.util.readFileSync.calls.length).to.eql(2);
-        validateCachedFiles(sharedIniFile, {
-          '/foo/bar/baz/.aws/credentials': {
-            'default': {
-              'aws_access_key_id': 'akid',
-              'aws_secret_access_key': 'secret',
-              'aws_session_token': 'session',
-            }
-          },
-          '/foo/bar/baz/.aws/config': {
-            'default': {
-              'aws_access_key_id': 'akid',
-              'aws_secret_access_key': 'secret',
-              'aws_session_token': 'session',
-            }
-          },
-        })
-      });
-  
-      it('load default profiles according to environmental variables', function() {
-        helpers.spyOn(os, 'homedir').andCallThrough();
-        helpers.spyOn(AWS.util, 'readFileSync').andReturn('');
-        process.env = {HOME: '/foo', USERPROFILE: '/bar', HOMEPATH: '/baz', HOMEDRIVE: '/quz'};
-        new SharedIniFile();
-        expect(AWS.util.readFileSync.calls.length).to.eql(2);
-        expect(AWS.util.readFileSync.calls[0].arguments).to.eql(['/foo/.aws/config']);
-        expect(AWS.util.readFileSync.calls[1].arguments).to.eql(['/foo/.aws/credentials']);
-        delete process.env.HOME;
-        new SharedIniFile();
-        expect(AWS.util.readFileSync.calls.length).to.eql(4);
-        expect(AWS.util.readFileSync.calls[2].arguments).to.eql(['/bar/.aws/config']);
-        expect(AWS.util.readFileSync.calls[3].arguments).to.eql(['/bar/.aws/credentials']);
-        delete process.env.USERPROFILE;
-        new SharedIniFile();
-        expect(AWS.util.readFileSync.calls.length).to.eql(6);
-        expect(AWS.util.readFileSync.calls[4].arguments).to.eql(['/quz/baz/.aws/config']);
-        expect(AWS.util.readFileSync.calls[5].arguments).to.eql(['/quz/baz/.aws/credentials']);
-        delete process.env.HOMEDRIVE;
-        new SharedIniFile();
-        expect(AWS.util.readFileSync.calls.length).to.eql(8);
-        expect(AWS.util.readFileSync.calls[6].arguments).to.eql(['C:/baz/.aws/config']);
-        expect(AWS.util.readFileSync.calls[7].arguments).to.eql(['C:/baz/.aws/credentials']);
-        expect(os.homedir.calls.length).to.eql(0);
+        expect(AWS.util.readFileSync.calls.length).to.equal(0);
       });
     });
   
@@ -156,6 +92,37 @@ if (AWS.util.isNode()) {
         sharedIniFile.loadFrom({filename: 'foo'});
         expect(AWS.util.readFileSync.calls.length).to.eql(1);
       });
+
+      it('load default profiles according to environmental variables', function() {
+        helpers.spyOn(os, 'homedir').andCallThrough();
+        helpers.spyOn(AWS.util, 'readFileSync').andReturn('');
+        process.env = {HOME: '/foo', USERPROFILE: '/bar', HOMEPATH: '/baz', HOMEDRIVE: '/quz'};
+        var iniFile = new SharedIniFile();
+        iniFile.loadFrom({isConfig: true});
+        iniFile.loadFrom({});
+        expect(AWS.util.readFileSync.calls.length).to.eql(2);
+        expect(AWS.util.readFileSync.calls[0].arguments).to.eql(['/foo/.aws/config']);
+        expect(AWS.util.readFileSync.calls[1].arguments).to.eql(['/foo/.aws/credentials']);
+        delete process.env.HOME;
+        iniFile.loadFrom({isConfig: true});
+        iniFile.loadFrom({});
+        expect(AWS.util.readFileSync.calls.length).to.eql(4);
+        expect(AWS.util.readFileSync.calls[2].arguments).to.eql(['/bar/.aws/config']);
+        expect(AWS.util.readFileSync.calls[3].arguments).to.eql(['/bar/.aws/credentials']);
+        delete process.env.USERPROFILE;
+        iniFile.loadFrom({isConfig: true});
+        iniFile.loadFrom({});
+        expect(AWS.util.readFileSync.calls.length).to.eql(6);
+        expect(AWS.util.readFileSync.calls[4].arguments).to.eql(['/quz/baz/.aws/config']);
+        expect(AWS.util.readFileSync.calls[5].arguments).to.eql(['/quz/baz/.aws/credentials']);
+        delete process.env.HOMEDRIVE;
+        iniFile.loadFrom({isConfig: true});
+        iniFile.loadFrom({});
+        expect(AWS.util.readFileSync.calls.length).to.eql(8);
+        expect(AWS.util.readFileSync.calls[6].arguments).to.eql(['C:/baz/.aws/config']);
+        expect(AWS.util.readFileSync.calls[7].arguments).to.eql(['C:/baz/.aws/credentials']);
+        expect(os.homedir.calls.length).to.eql(0);
+      });
     });
   
     describe('clearCachedFiles', function() {
@@ -163,16 +130,17 @@ if (AWS.util.isNode()) {
         process.env.HOME = '/foo';
         helpers.spyOn(AWS.util, 'readFileSync').andReturn('');
         var iniFile = new SharedIniFile();
+        iniFile.loadFrom();
         validateCachedFiles(iniFile, {
-          '/foo/.aws/config': {},
           '/foo/.aws/credentials': {},
-        })
-        iniFile.clearCachedFiles();
+        });
         helpers.spyOn(AWS.util, 'readFileSync').andReturn('');
-        iniFile.loadFrom({filename: '/foo/.aws/config'});
-        expect(Object.keys(iniFile.resolvedProfiles).length).to.eql(1);
-        iniFile.loadFrom({filename: '/foo/.aws/credentials'});
-        expect(Object.keys(iniFile.resolvedProfiles).length).to.eql(2);
+        expect(AWS.util.readFileSync.calls.length).to.eql(0);
+        iniFile.loadFrom();
+        expect(AWS.util.readFileSync.calls.length).to.eql(0);
+        iniFile.clearCachedFiles();
+        iniFile.loadFrom();
+        expect(AWS.util.readFileSync.calls.length).to.eql(1);
       })
     });
   });
